Share location data via a central server
Use websockets to host a central location server, and use
websocket-client to connect and share location data. The
server will broadcast location info for all players to all
connected clients. A future client addition might be an
additional window to display all other shared player names
and zones.

This code is currently a bit of a mess, as it was the
result of a proof-of-concept implementation. I still need
to clean it up a lot before making a PR.